+++ funciones.h
@@ -1,3 +1,3 @@
 float calcularMedia(int arr[], int size);
 void encontrar_moda(int numeros[], int tamano);
-void ordenar(int arr[], int size, int *mayor, int *menor, int *mediana);
+void ordenar(int arr[], int size, int *mayor, int *menor, float *mediana);
